refactor: split progress state test classes
