+++ src/Arius.Cli.Tests/ContentHashToPathTests.cs
@@ -1,0 +1,38 @@
+using Shouldly;
+
+namespace Arius.Cli.Tests;
+
+/// <summary>
+/// Verifies the <see cref="ProgressState.ContentHashToPath"/> reverse map is populated
+/// on hash and used for downstream events keyed by content hash.
+/// </summary>
+public class ContentHashToPathTests
+{
+    [Test]
+    public void ReverseMap_PopulatedOnHash_UsedForDownstreamEvents()
+    {
+        var state = new ProgressState();
+
+        state.AddFile("dir/file.bin", 500);
+        state.SetFileHashed("dir/file.bin", "aabbcc");
+
+        // Reverse map populated
+        state.ContentHashToPath.ContainsKey("aabbcc").ShouldBeTrue();
+        state.ContentHashToPath["aabbcc"].ShouldContain("dir/file.bin");
+
+        // Downstream event via reverse map: SetFileUploading transitions Hashed → Uploading
+        state.SetFileUploading("aabbcc");
+        state.TrackedFiles["dir/file.bin"].State.ShouldBe(FileState.Uploading);
+    }
+
+    [Test]
+    public void ReverseMap_PopulatedBeforeDownstreamEvents()
+    {
+        // FileHashedEvent sets both ContentHash and reverse map atomically
+        var state = new ProgressState();
+        state.AddFile("config.yml", 200);
+        state.SetFileHashed("config.yml", "xxyyzz");
+
+        state.ContentHashToPath.ContainsKey("xxyyzz").ShouldBeTrue();
+    }
+}
