+++ src/Arius.Cli.Tests/ProgressTests.cs
@@ -14,203 +14,6 @@ using Spectre.Console.Rendering;
 
 namespace Arius.Cli.Tests;
 
-// ── 7.1 TrackedFile state transitions: small file (tar) path ──────────────────
-
-/// <summary>
-/// Verifies <see cref="TrackedFile"/> state machine for the small-file/tar path:
-/// Hashing → Hashed → removed from TrackedFiles when added to TAR.
-/// </summary>
-public class TrackedFileSmallFilePathTests
-{
-    [Test]
-    public void SmallFilePath_StateTransitions_Correct()
-    {
-        var state = new ProgressState();
-
-        // FileHashingEvent → AddFile → State=Hashing
-        state.AddFile("notes.txt", 1024);
-        state.TrackedFiles.ContainsKey("notes.txt").ShouldBeTrue();
-        state.TrackedFiles["notes.txt"].State.ShouldBe(FileState.Hashing);
-
-        // FileHashedEvent → SetFileHashed → State=Hashed, reverse map populated
-        state.SetFileHashed("notes.txt", "def456");
-        state.TrackedFiles["notes.txt"].ContentHash.ShouldBe("def456");
-        state.TrackedFiles["notes.txt"].State.ShouldBe(FileState.Hashed);
-        state.ContentHashToPath["def456"].ShouldContain("notes.txt");
-        state.FilesHashed.ShouldBe(1L);
-
-        // TarEntryAddedEvent → RemoveFile (small file moves into TAR)
-        state.RemoveFile("notes.txt");
-        state.TrackedFiles.ContainsKey("notes.txt").ShouldBeFalse();
-    }
-}
-
-// ── 7.2 TrackedFile state transitions: large file path ────────────────────────
-
-/// <summary>
-/// Verifies <see cref="TrackedFile"/> state machine for the large-file/direct-upload path:
-/// Hashing → Hashed → Uploading → Done (removed).
-/// </summary>
-public class TrackedFileLargeFilePathTests
-{
-    [Test]
-    public void LargeFilePath_StateTransitions_Correct()
-    {
-        var state = new ProgressState();
-
-        // FileHashingEvent
-        state.AddFile("video.mp4", 5_000_000_000L);
-        state.TrackedFiles["video.mp4"].State.ShouldBe(FileState.Hashing);
-
-        // FileHashedEvent → State=Hashed (invisible)
-        state.SetFileHashed("video.mp4", "abc123");
-        state.TrackedFiles["video.mp4"].ContentHash.ShouldBe("abc123");
-        state.TrackedFiles["video.mp4"].State.ShouldBe(FileState.Hashed);
-
-        // ChunkUploadingEvent → SetFileUploading (only Hashed files promoted to Uploading)
-        state.SetFileUploading("abc123");
-        state.TrackedFiles["video.mp4"].State.ShouldBe(FileState.Uploading);
-
-        // ChunkUploadedEvent → RemoveFile
-        state.RemoveFile("video.mp4");
-        state.TrackedFiles.ContainsKey("video.mp4").ShouldBeFalse();
-    }
-
-    [Test]
-    public void SetFileUploading_OnlyPromotesHashedFiles()
-    {
-        // A file in Hashing state (not yet Hashed) should not transition to Uploading.
-        var state = new ProgressState();
-        state.AddFile("pending.txt", 100);
-        // Don't call SetFileHashed — still in Hashing state
-        state.TrackedFiles["pending.txt"].State.ShouldBe(FileState.Hashing);
-
-        // No ContentHashToPath entry yet, so SetFileUploading won't find it
-        state.SetFileUploading("nohash");
-        state.TrackedFiles["pending.txt"].State.ShouldBe(FileState.Hashing);
-    }
-}
-
-// ── ContentHashToPath reverse lookup ─────────────────────────────────────────
-
-/// <summary>
-/// Verifies the <see cref="ProgressState.ContentHashToPath"/> reverse map is populated
-/// on hash and used for downstream events keyed by content hash.
-/// </summary>
-public class ContentHashToPathTests
-{
-    [Test]
-    public void ReverseMap_PopulatedOnHash_UsedForDownstreamEvents()
-    {
-        var state = new ProgressState();
-
-        state.AddFile("dir/file.bin", 500);
-        state.SetFileHashed("dir/file.bin", "aabbcc");
-
-        // Reverse map populated
-        state.ContentHashToPath.ContainsKey("aabbcc").ShouldBeTrue();
-        state.ContentHashToPath["aabbcc"].ShouldContain("dir/file.bin");
-
-        // Downstream event via reverse map: SetFileUploading transitions Hashed → Uploading
-        state.SetFileUploading("aabbcc");
-        state.TrackedFiles["dir/file.bin"].State.ShouldBe(FileState.Uploading);
-    }
-
-    [Test]
-    public void ReverseMap_PopulatedBeforeDownstreamEvents()
-    {
-        // FileHashedEvent sets both ContentHash and reverse map atomically
-        var state = new ProgressState();
-        state.AddFile("config.yml", 200);
-        state.SetFileHashed("config.yml", "xxyyzz");
-
-        state.ContentHashToPath.ContainsKey("xxyyzz").ShouldBeTrue();
-    }
-}
-
-// ── ProgressState concurrent add/transition/remove ───────────────────────────
-
-/// <summary>
-/// Verifies <see cref="ProgressState"/> is thread-safe under concurrent operations.
-/// </summary>
-public class ProgressStateThreadSafetyTests
-{
-    [Test]
-    public async Task ConcurrentAddAndRemove_NoDataRaces()
-    {
-        var state = new ProgressState();
-        const int n = 5_000;
-
-        await Parallel.ForEachAsync(
-            Enumerable.Range(0, n),
-            new ParallelOptions { MaxDegreeOfParallelism = 8 },
-            (i, _) =>
-            {
-                var path = $"file{i}.bin";
-                var hash = $"hash{i:x8}";
-                state.AddFile(path, i * 1024L);
-                state.SetFileHashed(path, hash);
-                state.RemoveFile(path);
-                return ValueTask.CompletedTask;
-            });
-
-        // All removed — counter should be n (each SetFileHashed increments FilesHashed)
-        state.FilesHashed.ShouldBe(n);
-        state.TrackedFiles.Count.ShouldBe(0);
-    }
-
-    [Test]
-    public async Task ConcurrentIncrements_FilesHashed_CorrectTotal()
-    {
-        var state = new ProgressState();
-        const int n = 10_000;
-
-        for (var i = 0; i < n; i++)
-            state.AddFile($"file{i}", 100);
-
-        await Parallel.ForEachAsync(
-            Enumerable.Range(0, n),
-            new ParallelOptions { MaxDegreeOfParallelism = 8 },
-            (i, _) =>
-            {
-                state.SetFileHashed($"file{i}", $"hash{i:x8}");
-                return ValueTask.CompletedTask;
-            });
-
-        state.FilesHashed.ShouldBe(n);
-    }
-
-    [Test]
-    public async Task ConcurrentIncrements_FilesRestored_CorrectTotal()
-    {
-        var state = new ProgressState();
-        const int n = 8_000;
-
-        await Parallel.ForEachAsync(
-            Enumerable.Range(0, n),
-            new ParallelOptions { MaxDegreeOfParallelism = 8 },
-            (_, _) =>
-            {
-                state.IncrementFilesRestored(1024L);
-                return ValueTask.CompletedTask;
-            });
-
-        state.FilesRestored.ShouldBe(n);
-        state.BytesRestored.ShouldBe(n * 1024L);
-    }
-
-    [Test]
-    public void TotalFiles_NullUntilScanComplete()
-    {
-        var state = new ProgressState();
-        state.TotalFiles.ShouldBeNull();
-        state.SetScanComplete(1523, 1_000_000L);
-        state.TotalFiles.ShouldBe(1523L);
-        state.TotalBytes.ShouldBe(1_000_000L);
-        state.ScanComplete.ShouldBeTrue();
-    }
-}
-
 // ── 7.1 Archive notification handler unit tests ───────────────────────────────
 
 /// <summary>
@@ -535,107 +338,6 @@ public class NotificationHandlerTests
 
         state.RehydrationChunkCount.ShouldBe(7);
         state.RehydrationTotalBytes.ShouldBe(1024L * 1024L);
-    }
-}
-
-// ── TrackedTar lifecycle (7.3) ────────────────────────────────────────────────
-
-/// <summary>
-/// Verifies the full <see cref="TrackedTar"/> lifecycle:
-/// Accumulating → Sealing → Uploading → removed.
-/// </summary>
-public class TrackedTarLifecycleTests
-{
-    [Test]
-    public async Task TrackedTar_FullLifecycle_StateTransitions()
-    {
-        var state     = new ProgressState();
-        var startedH  = new TarBundleStartedHandler(state);
-        var entryH    = new TarEntryAddedHandler(state);
-        var sealingH  = new TarBundleSealingHandler(state);
-        var uploadingH = new ChunkUploadingHandler(state);
-        var uploadedH = new TarBundleUploadedHandler(state);
-
-        // Start tar
-        await startedH.Handle(new TarBundleStartedEvent(), CancellationToken.None);
-        state.TrackedTars[1].State.ShouldBe(TarState.Accumulating);
-
-        // Add files (need them in ContentHashToPath for entry handler)
-        var hashingH = new FileHashingHandler(state);
-        var hashedH  = new FileHashedHandler(state);
-        await hashingH.Handle(new FileHashingEvent("f1.txt", 100), CancellationToken.None);
-        await hashedH.Handle(new FileHashedEvent("f1.txt", "h1"), CancellationToken.None);
-        await hashingH.Handle(new FileHashingEvent("f2.txt", 200), CancellationToken.None);
-        await hashedH.Handle(new FileHashedEvent("f2.txt", "h2"), CancellationToken.None);
-
-        await entryH.Handle(new TarEntryAddedEvent("h1", 1, 100), CancellationToken.None);
-        await entryH.Handle(new TarEntryAddedEvent("h2", 2, 300), CancellationToken.None);
-
-        state.TrackedTars[1].FileCount.ShouldBe(2);
-        state.TrackedTars[1].AccumulatedBytes.ShouldBeGreaterThan(0);
-
-        // Seal
-        await sealingH.Handle(
-            new TarBundleSealingEvent(2, 300, "seal_hash", ["h1", "h2"]),
-            CancellationToken.None);
-        state.TrackedTars[1].State.ShouldBe(TarState.Sealing);
-        state.TrackedTars[1].TarHash.ShouldBe("seal_hash");
-        state.TrackedTars[1].TotalBytes.ShouldBe(300L);
-
-        // Upload starts
-        await uploadingH.Handle(new ChunkUploadingEvent("seal_hash", 300), CancellationToken.None);
-        state.TrackedTars[1].State.ShouldBe(TarState.Uploading);
-
-        // Upload complete → removed
-        await uploadedH.Handle(new TarBundleUploadedEvent("seal_hash", 200, 2), CancellationToken.None);
-        state.TrackedTars.ContainsKey(1).ShouldBeFalse();
-        state.TarsUploaded.ShouldBe(1L);
-    }
-}
-
-// ── 7.5 ChunkUploadingHandler: dual lookup ────────────────────────────────────
-
-/// <summary>
-/// Verifies the dual-lookup behavior of <see cref="ChunkUploadingHandler"/>:
-/// large files via <see cref="ProgressState.ContentHashToPath"/> and TAR bundles
-/// via <see cref="ProgressState.TrackedTars"/>.
-/// </summary>
-public class ChunkUploadingHandlerDualLookupTests
-{
-    [Test]
-    public async Task DualLookup_LargeFile_TakesPrecedence()
-    {
-        var state      = new ProgressState();
-        var hashingH   = new FileHashingHandler(state);
-        var hashedH    = new FileHashedHandler(state);
-        var uploadingH = new ChunkUploadingHandler(state);
-
-        await hashingH.Handle(new FileHashingEvent("big.bin", 10_000_000), CancellationToken.None);
-        await hashedH.Handle(new FileHashedEvent("big.bin", "largehash"), CancellationToken.None);
-
-        await uploadingH.Handle(new ChunkUploadingEvent("largehash", 10_000_000), CancellationToken.None);
-
-        state.TrackedFiles["big.bin"].State.ShouldBe(FileState.Uploading);
-        state.FilesUnique.ShouldBe(1L);
-    }
-
-    [Test]
-    public async Task DualLookup_TarBundle_TransitionsToUploading_NoFilesUniqueIncrement()
-    {
-        var state      = new ProgressState();
-        var startedH   = new TarBundleStartedHandler(state);
-        var sealingH   = new TarBundleSealingHandler(state);
-        var uploadingH = new ChunkUploadingHandler(state);
-
-        await startedH.Handle(new TarBundleStartedEvent(), CancellationToken.None);
-        await sealingH.Handle(
-            new TarBundleSealingEvent(1, 100, "tarhash99", ["h1"]),
-            CancellationToken.None);
-
-        await uploadingH.Handle(new ChunkUploadingEvent("tarhash99", 100), CancellationToken.None);
-
-        state.TrackedTars[1].State.ShouldBe(TarState.Uploading);
-        state.FilesUnique.ShouldBe(0L);  // TAR path does NOT increment FilesUnique
     }
 }
 
@@ -1359,36 +1061,6 @@ public class RestoreTcsCoordinationTests
     }
 }
 
-// ── TrackedFile BytesProcessed Interlocked update test ────────────────────────
-
-/// <summary>
-/// Verifies that <see cref="TrackedFile.SetBytesProcessed"/> reports the latest value
-/// under concurrent updates (last-writer semantics).
-/// </summary>
-public class TrackedFileBytesProcessedTests
-{
-    [Test]
-    public async Task BytesProcessed_UpdatesCorrectlyUnderContention()
-    {
-        var file = new TrackedFile("test.bin", 1_000_000L);
-
-        const int iterations = 10_000;
-        long      finalValue = 0;
-
-        await Parallel.ForEachAsync(
-            Enumerable.Range(1, iterations),
-            new ParallelOptions { MaxDegreeOfParallelism = 8 },
-            (i, _) =>
-            {
-                file.SetBytesProcessed(i);
-                Interlocked.Exchange(ref finalValue, i);
-                return ValueTask.CompletedTask;
-            });
-
-        file.BytesProcessed.ShouldBeInRange(1L, iterations);
-    }
-}
-
 // ── BuildArchiveDisplay: round-2 refinements ──────────────────────────────────
 
 /// <summary>
@@ -1507,93 +1179,6 @@ public class TruncateAndLeftJustifyTests
         var result = DisplayHelpers.TruncateAndLeftJustify("", 5);
         result.ShouldBe("     ");
         result.Length.ShouldBe(5);
-    }
-}
-
-// ── AddRestoreEvent ring buffer ───────────────────────────────────────────────
-
-/// <summary>
-/// Verifies <see cref="ProgressState.AddRestoreEvent"/> caps the queue at 10 entries.
-/// </summary>
-public class AddRestoreEventTests
-{
-    [Test]
-    public void AddRestoreEvent_CapAt10_KeepsMostRecent()
-    {
-        var state = new ProgressState();
-
-        for (var i = 1; i <= 15; i++)
-            state.AddRestoreEvent($"file{i}.txt", i * 100L, skipped: false);
-
-        state.RecentRestoreEvents.Count.ShouldBe(10);
-
-        var paths = state.RecentRestoreEvents.Select(e => e.RelativePath).ToList();
-        paths.ShouldContain("file15.txt");
-        paths.ShouldContain("file6.txt");
-        paths.ShouldNotContain("file5.txt");
-        paths.ShouldNotContain("file1.txt");
-    }
-
-    [Test]
-    public void AddRestoreEvent_BelowCap_AllRetained()
-    {
-        var state = new ProgressState();
-
-        for (var i = 1; i <= 5; i++)
-            state.AddRestoreEvent($"file{i}.txt", 100L, skipped: i % 2 == 0);
-
-        state.RecentRestoreEvents.Count.ShouldBe(5);
-    }
-}
-
-// ── IncrementFilesRestored / IncrementFilesSkipped byte accumulators ──────────
-
-/// <summary>
-/// Verifies byte accumulators are updated alongside counters.
-/// </summary>
-public class RestoreByteAccumulatorTests
-{
-    [Test]
-    public void IncrementFilesRestored_UpdatesCounterAndBytes()
-    {
-        var state = new ProgressState();
-
-        state.IncrementFilesRestored(1000L);
-        state.IncrementFilesRestored(2000L);
-
-        state.FilesRestored.ShouldBe(2L);
-        state.BytesRestored.ShouldBe(3000L);
-    }
-
-    [Test]
-    public void IncrementFilesSkipped_UpdatesCounterAndBytes()
-    {
-        var state = new ProgressState();
-
-        state.IncrementFilesSkipped(512L);
-        state.IncrementFilesSkipped(512L);
-
-        state.FilesSkipped.ShouldBe(2L);
-        state.BytesSkipped.ShouldBe(1024L);
-    }
-}
-
-// ── SetRehydration sets both fields ──────────────────────────────────────────
-
-/// <summary>
-/// Verifies <see cref="ProgressState.SetRehydration"/> sets both chunk count and byte total.
-/// </summary>
-public class SetRehydrationTests
-{
-    [Test]
-    public void SetRehydration_SetsBothFields()
-    {
-        var state = new ProgressState();
-
-        state.SetRehydration(5, 10_485_760L);
-
-        state.RehydrationChunkCount.ShouldBe(5);
-        state.RehydrationTotalBytes.ShouldBe(10_485_760L);
     }
 }
 
@@ -1865,97 +1450,6 @@ public class BuildRestoreDisplayTests
     }
 }
 
-// ── 8.1 TrackedDownload lifecycle ─────────────────────────────────────────────
-
-/// <summary>
-/// Verifies <see cref="TrackedDownload"/> lifecycle: add, update bytes, remove on completion.
-/// Covers large file (removed by FileRestoredHandler) and tar bundle (removed by ChunkDownloadCompletedHandler).
-/// </summary>
-public class TrackedDownloadLifecycleTests
-{
-    [Test]
-    public async Task LargeFile_TrackedDownload_AddUpdateRemove()
-    {
-        var state = new ProgressState();
-
-        // Simulate CreateDownloadProgress adding a TrackedDownload for a large file
-        // Key is RelativePath (the identifier passed from RestoreCommandHandler for large files)
-        var td = new TrackedDownload("photos/sunset.jpg", DownloadKind.LargeFile, "photos/sunset.jpg", compressedSize: 25_400_000, originalSize: 50_000_000);
-        state.TrackedDownloads.TryAdd("photos/sunset.jpg", td);
-
-        state.TrackedDownloads.Count.ShouldBe(1);
-        state.TrackedDownloads["photos/sunset.jpg"].DisplayName.ShouldBe("photos/sunset.jpg");
-        state.TrackedDownloads["photos/sunset.jpg"].Kind.ShouldBe(DownloadKind.LargeFile);
-
-        // Simulate byte-level progress updates
-        td.SetBytesDownloaded(12_300_000);
-        td.BytesDownloaded.ShouldBe(12_300_000L);
-
-        td.SetBytesDownloaded(25_400_000);
-        td.BytesDownloaded.ShouldBe(25_400_000L);
-
-        // FileRestoredHandler should remove the TrackedDownload by direct key lookup on RelativePath
-        var handler = new FileRestoredHandler(state);
-        await handler.Handle(new FileRestoredEvent("photos/sunset.jpg", 50_000_000L), CancellationToken.None);
-
-        state.TrackedDownloads.ContainsKey("photos/sunset.jpg").ShouldBeFalse("TrackedDownload should be removed after FileRestoredEvent");
-        state.RestoreBytesDownloaded.ShouldBe(25_400_000L, "RestoreBytesDownloaded should be incremented by CompressedSize");
-        state.FilesRestored.ShouldBe(1L);
-        state.BytesRestored.ShouldBe(50_000_000L);
-    }
-
-    [Test]
-    public async Task TarBundle_TrackedDownload_AddUpdateRemove()
-    {
-        var state = new ProgressState();
-
-        // Simulate CreateDownloadProgress adding a TrackedDownload for a tar bundle
-        var td = new TrackedDownload("ab12cd34", DownloadKind.TarBundle, "TAR bundle (3 files, 847 KB)", compressedSize: 15_200_000, originalSize: 847_000);
-        state.TrackedDownloads.TryAdd("ab12cd34", td);
-
-        state.TrackedDownloads.Count.ShouldBe(1);
-        state.TrackedDownloads["ab12cd34"].Kind.ShouldBe(DownloadKind.TarBundle);
-
-        // Simulate byte-level progress
-        td.SetBytesDownloaded(4_800_000);
-        td.BytesDownloaded.ShouldBe(4_800_000L);
-
-        // ChunkDownloadCompletedHandler should remove the TrackedDownload
-        var handler = new ChunkDownloadCompletedHandler(state);
-        await handler.Handle(new ChunkDownloadCompletedEvent("ab12cd34", 3, 15_200_000), CancellationToken.None);
-
-        state.TrackedDownloads.ContainsKey("ab12cd34").ShouldBeFalse("TrackedDownload should be removed after ChunkDownloadCompletedEvent");
-        state.RestoreBytesDownloaded.ShouldBe(15_200_000L, "RestoreBytesDownloaded should be incremented by CompressedSize");
-    }
-
-    [Test]
-    public void TrackedDownload_BytesDownloaded_InterlockedUpdate()
-    {
-        var td = new TrackedDownload("key1", DownloadKind.LargeFile, "file.bin", 1_000_000, 2_000_000);
-
-        td.BytesDownloaded.ShouldBe(0L);
-        td.SetBytesDownloaded(500_000);
-        td.BytesDownloaded.ShouldBe(500_000L);
-        td.SetBytesDownloaded(1_000_000);
-        td.BytesDownloaded.ShouldBe(1_000_000L);
-    }
-
-    [Test]
-    public async Task FileRestoredHandler_NoTrackedDownload_StillUpdatesCounters()
-    {
-        // When no TrackedDownload exists (e.g. file from tar bundle), handler should still work
-        var state   = new ProgressState();
-        var handler = new FileRestoredHandler(state);
-
-        await handler.Handle(new FileRestoredEvent("some/file.txt", 1024L), CancellationToken.None);
-
-        state.FilesRestored.ShouldBe(1L);
-        state.BytesRestored.ShouldBe(1024L);
-        // No TrackedDownload removal expected, RestoreBytesDownloaded stays at 0
-        state.RestoreBytesDownloaded.ShouldBe(0L);
-    }
-}
-
 // ── 8.2 TreeTraversalProgressHandler ──────────────────────────────────────────
 
 /// <summary>
@@ -2090,43 +1584,6 @@ public class ChunkResolutionCompleteHandlerByteTotalsTests
         state.TarChunkCount.ShouldBe(5);
         state.RestoreTotalOriginalSize.ShouldBe(500_000_000L);
         state.RestoreTotalCompressedBytes.ShouldBe(200_000_000L);
-    }
-}
-
-// ── 8.11 Thread safety: concurrent TrackedDownload add/update/remove ──────────
-
-/// <summary>
-/// Verifies thread safety of concurrent TrackedDownload add/update/remove from 4 workers.
-/// </summary>
-public class TrackedDownloadThreadSafetyTests
-{
-    [Test]
-    public async Task ConcurrentAddUpdateRemove_NoDataRaces()
-    {
-        var state = new ProgressState();
-        const int n = 1_000;
-
-        await Parallel.ForEachAsync(
-            Enumerable.Range(0, n),
-            new ParallelOptions { MaxDegreeOfParallelism = 4 },
-            (i, ct) =>
-            {
-                var key = $"file_{i}.bin"; // Key is RelativePath for large files
-                var td = new TrackedDownload(key, DownloadKind.LargeFile, $"file_{i}.bin", 1_000_000, 2_000_000);
-                state.TrackedDownloads.TryAdd(key, td);
-
-                // Simulate byte-level progress
-                td.SetBytesDownloaded(500_000);
-                td.SetBytesDownloaded(1_000_000);
-
-                // Remove on completion
-                state.TrackedDownloads.TryRemove(key, out _);
-                state.AddRestoreBytesDownloaded(1_000_000);
-                return ValueTask.CompletedTask;
-            });
-
-        state.TrackedDownloads.Count.ShouldBe(0, "All tracked downloads should be removed");
-        state.RestoreBytesDownloaded.ShouldBe(n * 1_000_000L, "All bytes should be accounted for");
     }
 }
