+++ src/Arius.Cli.Tests/TrackedFileLargeFilePathTests.cs
@@ -1,0 +1,47 @@
+using Shouldly;
+
+namespace Arius.Cli.Tests;
+
+/// <summary>
+/// Verifies <see cref="TrackedFile"/> state machine for the large-file/direct-upload path:
+/// Hashing → Hashed → Uploading → Done (removed).
+/// </summary>
+public class TrackedFileLargeFilePathTests
+{
+    [Test]
+    public void LargeFilePath_StateTransitions_Correct()
+    {
+        var state = new ProgressState();
+
+        // FileHashingEvent
+        state.AddFile("video.mp4", 5_000_000_000L);
+        state.TrackedFiles["video.mp4"].State.ShouldBe(FileState.Hashing);
+
+        // FileHashedEvent → State=Hashed (invisible)
+        state.SetFileHashed("video.mp4", "abc123");
+        state.TrackedFiles["video.mp4"].ContentHash.ShouldBe("abc123");
+        state.TrackedFiles["video.mp4"].State.ShouldBe(FileState.Hashed);
+
+        // ChunkUploadingEvent → SetFileUploading (only Hashed files promoted to Uploading)
+        state.SetFileUploading("abc123");
+        state.TrackedFiles["video.mp4"].State.ShouldBe(FileState.Uploading);
+
+        // ChunkUploadedEvent → RemoveFile
+        state.RemoveFile("video.mp4");
+        state.TrackedFiles.ContainsKey("video.mp4").ShouldBeFalse();
+    }
+
+    [Test]
+    public void SetFileUploading_OnlyPromotesHashedFiles()
+    {
+        // A file in Hashing state (not yet Hashed) should not transition to Uploading.
+        var state = new ProgressState();
+        state.AddFile("pending.txt", 100);
+        // Don't call SetFileHashed — still in Hashing state
+        state.TrackedFiles["pending.txt"].State.ShouldBe(FileState.Hashing);
+
+        // No ContentHashToPath entry yet, so SetFileUploading won't find it
+        state.SetFileUploading("nohash");
+        state.TrackedFiles["pending.txt"].State.ShouldBe(FileState.Hashing);
+    }
+}
